renaming ./examples/libs to ./examples/common
