+++ examples/common/asset_downloader.py
